requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
dj-database-url==3.1.0
Django==6.0
django-mathfilters==1.0.0
django-widget-tweaks==1.5.1
gunicorn==23.0.0
idna==3.11
packaging==25.0
pillow==12.1.0
psycopg==3.3.2
psycopg-binary==3.3.2
psycopg2-binary==2.9.11
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.5
stripe==14.1.0
typing_extensions==4.15.0
urllib3==2.6.2
whitenoise==6.11.0
